+++ modules/ApiClient.ts
@@ -1,7 +1,7 @@
 'use strict';
 
 import Ajax from './Ajax';
-import { RegisterParams, AdsFilters, LoginParams } from './Types';
+import { RegisterParams, AdsFilters, LoginParams, EditParams } from './Types';
 
 const APIClient = {
     BASE_URL: `http://${window.location.hostname}:8008/api`,
@@ -80,6 +80,49 @@ const APIClient = {
 
         return Ajax.post({ url, body });
     },
+
+    async city (name: string): Promise<any>  {
+        const url = this.BASE_URL + `/getPlacesPerCity/${name}`;
+        return Ajax.get(url);
+    },
+
+    async profile (): Promise<any> {
+        const url = this.BASE_URL + '/getUserById';
+        return Ajax.get(url);
+    },
+
+    async editProfile ({
+        username,
+        name,
+        email,
+        sex,
+        address,
+        birthdate,
+        isHost,
+        password,
+        avatar,
+    }: EditParams): Promise<any> {
+        const url = this.BASE_URL + '/putUser';
+        const formData = new FormData();
+    
+        const metadata = {
+            username: username,
+            name: name,
+            email: email,
+            sex: sex,
+            address: address,
+            birthdate: birthdate,
+            isHost: isHost,
+            password: password,
+        };
+    
+        formData.append('metadata', JSON.stringify(metadata));
+        if (avatar){
+            formData.append('avatar', avatar);
+        }
+    
+        return Ajax.put({url, body: formData});
+    }
 };
 
 export default APIClient;
